% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy-measures.R
\name{accuracy_measures}
\alias{accuracy_measures}
\title{Accuracy measures}
\usage{
accuracy_measures(object, y, yhat, average_folds = TRUE)
}
\arguments{
\item{object}{an object containing cross-validation folds and predictions.}

\item{y}{string. Name of actual column.}

\item{yhat}{string. Name of prediction column.}

\item{average_folds}{logical. Should accuracy measures be averaged across all
folds? If \code{TRUE} one line is returned. If \code{FALSE} the accuracy measures are
returned for each cross-validation fold.}
}
\value{
A data frame with the accuracy measures listed above.
}
\description{
Returns prediction accuracy measures for:
}
\details{
\itemize{
\item Absolute error (AE)
\item Mean absolute error (MAE)
\item Mean absolute percentage error (MAPE)
\item Root mean square error (RMSE)
\item Symmetric mean absolute percentage error (SMAPE1 and SMAPE2 - see
discussion)
}
}
\examples{
kfold_data <- kfold_cv(iris)
# TODO need to figure out how to do the prediction function, might need to be something the user codes since each prediction method is different.
kfold_data <- predict(kfold_data)
accuracy_measures(kfold_data, y = "Actual", yhat = "Prediction")
}
