% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.validatr}
\alias{autoplot.validatr}
\title{Plot accuracy measure densities}
\usage{
\method{autoplot}{validatr}(object, ...)
}
\arguments{
\item{object}{a validatr object produced by \code{assess()}.}

\item{...}{other arguments passed.}
}
\value{
A ggplot object.
}
\description{
Produces violin plots for accuracy measures. Jittered points are also used
to show actual values.
}
\examples{
iris \%>\%
  validatr(y = Sepal.Length, k = 3) \%>\%
  model(LM1 = lm(Sepal.Length ~ ., data = train),
        LM2 = lm(Sepal.Length ~ Sepal.Width + Petal.Width, data = train)) \%>\%
  predict(LM1 = predict(LM1, newdata = validation),
          LM2 = predict(LM2, newdata = validation)) \%>\%
  assess() \%>\%
  autoplot()
}
