% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.validatr_accuracy}
\alias{autoplot.validatr_accuracy}
\title{Plot accuracy measure densities}
\usage{
autoplot.validatr_accuracy(object, ...)
}
\arguments{
\item{object}{a data frame produced by \code{calc_accuracy()}.}

\item{...}{}
}
\value{
A ggplot object.
}
\description{
Plot accuracy measure densities
}
\examples{
validatr(iris, k = 3) \%>\%
  fit_models(Model1 = lm(Sepal.Length ~ ., data = train),
             Model2 = lm(Sepal.Length ~ Sepal.Width + Petal.Width, data = train)) \%>\%
  calc_predictions(Model1 = predict(Model1, newdata = validation),
                   Model2 = predict(Model2, newdata = validation)) \%>\%
  calc_accuracy(y = "Sepal.Length", average_folds = FALSE) \%>\%
  autoplot()
}
