% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-accuracy.R
\name{calc_accuracy}
\alias{calc_accuracy}
\title{Accuracy measures}
\usage{
calc_accuracy(.object, y)
}
\arguments{
\item{.object}{a \code{validatr} object containing cross-validation folds and predictions.}

\item{y}{string. Name of actual column.}
}
\value{
A data frame with the accuracy measures listed above.
}
\description{
Returns prediction accuracy measures.
}
\details{
Accuracy measures returned include:
\itemize{
\item Absolute error (AE)
\item Mean absolute error (MAE)
\item Mean absolute percentage error (MAPE)
\item Root mean square error (RMSE)
\item Symmetric mean absolute percentage error (SMAPE)
\item Mean absolute scaled error (MASE)
}
}
\examples{
validatr(iris, k = 3) \%>\%
  fit_models(Model1 = lm(Sepal.Length ~ ., data = train),
             Model2 = lm(Sepal.Length ~ Sepal.Width + Petal.Width, data = train)) \%>\%
  calc_predictions(Model1 = predict(Model1, newdata = validation),
                   Model2 = predict(Model2, newdata = validation)) \%>\%
  calc_accuracy(y = "Sepal.Length")
}
