% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-models.R
\name{fit_models}
\alias{fit_models}
\title{Fit models to training sets}
\usage{
fit_models(.object, ...)
}
\arguments{
\item{.object}{a validatr object produced using \code{validatr()}.}

\item{...}{}
}
\value{
A validatr object with models fitted for each fold.
}
\description{
Fits specified models to training sets.
}
\examples{

validatr(iris, k = 3) \%>\%
  fit_models(Model1 = lm(Sepal.Length ~ ., data = train),
             Model2 = lm(Sepal.Length ~ Sepal.Width + Petal.Width, data = train))
}
