% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold-cv.R
\name{kfold_cv}
\alias{kfold_cv}
\title{K-fold cross validation}
\usage{
kfold_cv(data, k = 10)
}
\arguments{
\item{data}{a data frame containing relevant variables for modelling.}

\item{k}{integer. Number of folds.}
}
\value{
\code{kfold_cv} returns a list of train and validation data frames. Each element of the
list corresponds to a fold.
}
\description{
\strong{TODO: Change to get_folds(data, method = "kfold-cv", k = 10)}
}
\details{
Performs k-fold cross validation. A list with k elements is returned. Each
element contains a train and validation element, both of which are data frames.
}
\examples{

kfold_data <- kfold_cv(iris, k = 5)
head(kfold_data[[5]]$train)
head(kfold_data[[5]]$validation)
}
