% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{pinball_loss}
\alias{pinball_loss}
\title{Pinball loss function}
\usage{
pinball_loss(tau, y, q)
}
\arguments{
\item{tau}{a vector of integers giving the quantile to calculate the pinball
loss score for.}

\item{y}{a numeric vector of actual values.}

\item{q}{a numeric vector of predicted values for quantile \code{tau}.}
}
\value{
Pinball loss score.
}
\description{
Calculates the pinball loss score for a given quantile.
}
