% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatr.R
\name{validatr}
\alias{validatr}
\title{Initialise validatr}
\usage{
validatr(data, y, k = 10, ts = NULL, start = NULL, horizon = NULL,
  shift = NULL)
}
\arguments{
\item{data}{data frame containing variables for modelling.}

\item{y}{dependent variable name. Non-standard evaluation.}

\item{k}{integer. Number of folds.}

\item{ts}{time-series variable name. Non-standard evaluation.}

\item{start}{numeric, date or POSIX object specifying the start date for
time-series validation folds.}

\item{horizon}{forecast horizon to evaluate.}

\item{shift}{length of time to move forward for each new fold.}
}
\value{
\code{validatr} returns an initial validatr object. This object contains cross
validation folds and validation parameters.
}
\description{
Initialises a validatr object.
}
\details{
The type of data being tested influences how the validatr and it's methods
respond. The following types are supported:
\itemize{
\item \strong{regression} regression data (default).
\item \strong{ts} time-series data (\code{ts} argument specified).
\item \strong{classification} classification data (\code{y} variable character, factor or
logical).
}

The type of cross-validation and accuracy measures to be calculated are
influenced by this parameter. For regression, k-fold cross-validation is
carried out and requires the number of folds \code{k} to be specified. Leave one
out cross-validation can easily be accomplished by setting \code{k} equal to the
number of observations.

For time-series, time-series cross-validation takes place. This requires the
\code{start}, \code{horizon}, \code{shift} and \code{ts} arguments to be specified:
\itemize{
\item \code{start} is the start of the first fold.
\item \code{horizon} is the length of the fold.
\item \code{shift} is the length of time to move forward.
\item \code{ts} is the name of the variable containing time-series data.
}

If \code{start} is numeric, then \code{horizon} and \code{shift} are also numeric. If
\code{start} is date or POSIX, then \code{horizon} and \code{shift} follow the same
convention as for \code{seq.Date} and \code{seq.POSIXt}. Hence, they are a character
string, containing one of "sec", "min", "hour", "day", "DSTday", "week",
"month", "quarter" or "year".

Finally, classification carries out k-fold cross validation as well, but its
accruacy measures will be different to regression.
}
\examples{

validatr_obj <- validatr(iris, y = Sepal.Length, k = 5)
head(validatr_obj$folds[[5]]$train)
head(validatr_obj$folds[[5]]$validation)
}
