% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot}
\alias{autoplot}
\title{Plot accuracy measure densities}
\usage{
autoplot(object, ...)
}
\arguments{
\item{object}{a validatr object produced by \code{assess()}.}

\item{...}{}
}
\value{
A ggplot object.
}
\description{
Plot accuracy measure densities
}
\examples{
validatr(iris, k = 3) \%>\%
  model(LM1 = lm(Sepal.Length ~ ., data = train),
        LM2 = lm(Sepal.Length ~ Sepal.Width + Petal.Width, data = train)) \%>\%
  predict(LM1 = predict(LM1, newdata = validation),
          LM2 = predict(LM2, newdata = validation)) \%>\%
  assess(y = "Sepal.Length") \%>\%
  autoplot()
}
