% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot}
\alias{autoplot}
\title{Plot accuracy measure densities}
\usage{
autoplot(object, ...)
}
\arguments{
\item{object}{a data frame produced by \code{calc_accuracy()}.}

\item{...}{}
}
\value{
A ggplot object.
}
\description{
Plot accuracy measure densities
}
\examples{
validatr(iris, k = 3) \%>\%
  fit_models(LM1 = lm(Sepal.Length ~ ., data = train),
             LM2 = lm(Sepal.Length ~ Sepal.Width + Petal.Width, data = train)) \%>\%
  calc_predictions(LM1 = predict(LM1, newdata = validation),
                   LM2 = predict(LM2, newdata = validation)) \%>\%
  calc_accuracy(y = "Sepal.Length", average_folds = FALSE) \%>\%
  autoplot()
}
