% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy-measures.R
\name{calc_accuracy}
\alias{calc_accuracy}
\title{Accuracy measures}
\usage{
calc_accuracy(.object, y, yhat, average_folds = TRUE)
}
\arguments{
\item{.object}{a \code{validatr} object containing cross-validation folds and predictions.}

\item{y}{string. Name of actual column.}

\item{yhat}{string. Name of prediction column.}

\item{average_folds}{logical. Should accuracy measures be averaged across all
folds? If \code{TRUE} one line is returned. If \code{FALSE} the accuracy measures are
returned for each cross-validation fold.}
}
\value{
A data frame with the accuracy measures listed above.
}
\description{
Returns prediction accuracy measures.
}
\details{
Accuracy measures returned include:
\itemize{
\item Absolute error (AE)
\item Mean absolute error (MAE)
\item Mean absolute percentage error (MAPE)
\item Root mean square error (RMSE)
\item Symmetric mean absolute percentage error (SMAPE1 and SMAPE2 - see
discussion)
}
}
\examples{
kfold_cv(iris, k = 3) \%>\%
  fit_models("Model1 = lm(Sepal.Length ~ ., data = train)",
             "Model2 = lm(Sepal.Length ~ Sepal.Width + Petal.Width, data = train)") \%>\%
  calc_predictions("Prediction1 = predict(Model1, newdata = test)",
                   "Prediction2 = predict(Model2, newdata = test)") \%>\%
  calc_accuracy(y = "Sepal.Length",
                    yhat = c("Prediction1", "Prediction2"))
}
