% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{Fit models to training sets}
\usage{
fit_models(.data, ...)
}
\arguments{
\item{...}{}
}
\description{
Fits specified models to training sets.
}
\examples{

kfold_cv(iris, k = 3) \%>\%
  fit_models(Model1 = lm(Sepal.Length ~ ., data = train),
             Model2 = lm(Sepal.Length ~ Sepal.Width + Petal.Width, data = train))
}
