% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.validatr}
\alias{predict.validatr}
\title{Predict on validation set}
\usage{
\method{predict}{validatr}(object, ...)
}
\arguments{
\item{object}{a \code{validatr} object produced by \code{model()}.}

\item{...}{strings specifying what predictions are needed. Each string, if
evaluated, should return a vector of predictions.}
}
\value{
A \code{validatr} object with predictions element.
}
\description{
Carries out prediction on validation set for each model.
}
\examples{
iris \%>\%
  validatr(Sepal.Length, k = 3) \%>\%
  model(Model1 = lm(Sepal.Length ~ ., data = train),
        Model2 = lm(Sepal.Length ~ Sepal.Width + Petal.Width, data = train)) \%>\%
  predict(Model1 = predict(Model1, newdata = validation),
          Model2 = predict(Model2, newdata = validation))
}
